agent_results/end2end_gemini/iclr2025_mldpr/claude_code/requirements.txt
numpy>=1.20.0
pandas>=1.3.0
matplotlib>=3.4.0
seaborn>=0.11.0
torch>=1.9.0
torchvision>=0.10.0
scikit-learn>=0.24.0
tqdm>=4.62.0
datasets>=1.18.0
transformers>=4.18.0
pillow>=8.3.0
deap>=1.3.1  # For evolutionary algorithms
tensorboard>=2.6.0
pyyaml>=6.0.0
joblib>=1.1.0
opencv-python>=4.5.0
scipy>=1.7.0